backend/scripts/requirements.txt
# Core dependencies
requests>=2.31.0
python-dotenv>=1.0.0

# PDF processing
PyMuPDF>=1.23.0
Pillow>=10.0.0

# AI/ML
google-generativeai>=0.3.0
langchain>=0.1.0
langchain-community>=0.0.20
sentence-transformers>=2.2.0
chromadb>=0.4.0

# Web scraping
playwright>=1.40.0
selenium>=4.15.0

# Utilities
beautifulsoup4>=4.12.0
lxml>=4.9.0
urllib3>=2.0.0

# Optional: For better performance
numpy>=1.24.0
pandas>=2.0.0

# Development (optional)
pytest>=7.4.0
black>=23.0.0
flake8>=6.0.0 